students/k3343/Mordovkin_Maxim/Lr3/requirements.txt
fastapi~=0.115.12
uvicorn
sqlmodel~=0.0.24
psycopg2-binary
alembic~=1.16.1
python-dotenv~=1.1.0
PyJWT~=2.10.1
python-multipart
pydantic~=2.11.5
SQLAlchemy~=2.0.41
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.13.4
asyncpg~=0.30.0
aiohttp~=3.11.18
celery==5.3.5
redis==4.5.5